Μετασχηματισα το: plan -->Plan service -->PlanImpl service -->Dao interface -->Dao implementor

+++ src/main/java/fashion/daos/PayMethodDao.java
@@ -4,5 +4,5 @@ import fashion.entity.PayMethod;
 
 
 public interface PayMethodDao {
-    public PayMethod findByPaymentMethod(int id);
+    public PayMethod findByPaymentMethod(short id);
 }
