+++ src/main/java/fashion/daos/SubscriptionStatuDaoImpl.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 public class SubscriptionStatuDaoImpl extends SuperDao implements SubscriptionStatuDao{
 
     @Override
-    public SubscriptionStatu findById(int id) {
+    public SubscriptionStatu findById(short id) {
         return(getSession().createNamedQuery("SubscriptionStatu.findById", SubscriptionStatu.class).setParameter("id", id).getSingleResult());
     }
